Add some more type, monoid and monad notation
Except for 'return' and 'join', all of them already used in

  http://hackage.haskell.org/package/base-unicode-symbols

Eta and Mu are common notations for 'return' and 'join', e.g.

  http://www.euclideanspace.com/maths/discrete/category/higher/monad/index.htm

Even though they are letter-functions (similar to as we use lambda symbol
for '\'), they can be easily made distinguishable from letter-variables by
the colorscheme (using colors or bold type, alike to vectors vs. variables
in math.)